- move `UnspentUnexpiredTickets()` to tickets.go - make httpGet public and move to utils.go - cleanup comments
