+++ wallet.go
@@ -289,18 +289,3 @@ func (wallet *Wallet) DecryptSeed(privatePassphrase []byte) (string, error) {
 
 	return decryptWalletSeed(privatePassphrase, wallet.EncryptedSeed)
 }
-
-// UnspentUnexpiredTickets returns all Unmined, Immature and Live tickets.
-func (wallet *Wallet) UnspentUnexpiredTickets() ([]Transaction, error) {
-	var tickets []Transaction
-	for _, filter := range []int32{TxFilterUnmined, TxFilterImmature, TxFilterLive} {
-		tx, err := wallet.GetTransactionsRaw(0, 0, filter, true)
-		if err != nil {
-			return nil, err
-		}
-
-		tickets = append(tickets, tx...)
-	}
-
-	return tickets, nil
-}
